refactor(PaintingModule) : move logic to service layer
- allow reuse logic in other services
